+++ markymark/conf.py
@@ -1,20 +1,21 @@
 from django.conf import settings
+from . import defaults
 
 
-MARKYMARK_EXTENSIONS = getattr(settings, 'MARKYMARK_EXTENSIONS', [
-    'markymark.extensions.autolink',
-])
+MARKYMARK_EXTENSIONS = getattr(
+    settings,
+    'MARKYMARK_EXTENSIONS',
+    defaults.MARKYMARK_EXTENSIONS
+)
 
+MARKYMARK_CSS = getattr(
+    settings,
+    'MARKYMARK_CSS',
+    defaults.MARKYMARK_CSS
+)
 
-MARKYMARK_CSS = getattr(settings, 'MARKYMARK_CSS', [
-    'markdown/css/markdown-editor.css',
-    'markdown/css/markdown-editor-adminfix.css',
-])
-
-
-MARKYMARK_JS = getattr(settings, 'MARKYMARK_JS', [
-    'markdown/js/markdown-init.js',
-    'markdown/js/markdown.js',
-    'markdown/js/markdown-editor.js',
-    'markdown/js/plugins/clean.js'
-])
+MARKYMARK_JS = getattr(
+    settings,
+    'MARKYMARK_JS',
+    defaults.MARKYMARK_JS
+)