+++ markymark/defaults.py
@@ -1,0 +1,17 @@
+MARKYMARK_EXTENSIONS = [
+    'markymark.extensions.autolink',
+]
+
+
+MARKYMARK_CSS = [
+    'markdown/css/markdown-editor.css',
+    'markdown/css/markdown-editor-adminfix.css',
+]
+
+
+MARKYMARK_JS = [
+    'markdown/js/markdown-init.js',
+    'markdown/js/markdown.js',
+    'markdown/js/markdown-editor.js',
+    'markdown/js/plugins/clean.js'
+]